+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Fix pace computation when it is under an hour
 
+### Changed
+
+- Retrieve LTI Consumer plugin context through an API endpoint
+
 ## [2.3.3] - 2021-03-25
 
 ### Fixed